requirements.txt
openai #for communicating with the llm
python-dotenv #for loading the env variables
gradio #for the ui
requests
PyPDF2
